requirements.txt
annotated-types==0.5.0
anyio==3.7.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
fastapi==0.103.0
Flask==2.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
Werkzeug==2.3.7
